# Conflicts:
#	src/main/java/ru/eventorg/controller/ParticipantsListController.java
#	src/main/java/ru/eventorg/service/EventService.java
#	src/main/java/ru/eventorg/service/ParticipantsListService.java

+++ src/main/java/ru/eventorg/entity/DebtEntity.java
@@ -20,6 +20,6 @@ public class DebtEntity {
     private String payerId;
     private String recipientId;
     private Integer statusId;
-    private BigDecimal debtAmount;
+    private Float debtAmount;
     private Integer eventId;
 }
